cdm::register::sp

Macro read

macro_rules! read {
    () => { ... };
}

Reads the value of the stack pointer (SP) register.

Requirements

The macro expands to an inline assembly block. It requires enabling the asm_experimental_arch compiler feature in your crate. This can be done by adding the following line to the top of your crate’s main.rs or lib.rs:

#![feature(asm_experimental_arch)]

Note: this is a macro and not a function, because SP is often modified in the function prologue, so a function may return an inaccurate result, pointing to the end of the next stack frame instead of the frame of the function where it was used. The function could (and most likely would) be inlined, but this would be unreliable, because even #[inline(always)] does not guarantee this.